refactor header into server component
